getmatchData confirm

+++ getmatchData.py
@@ -18,9 +18,11 @@ for player in players:
     print(player)
     matchs = pd.read_csv(f"player/{player}", header = 0 , index_col = 0)
     try:
-        for matchId in matchs["gameId"].iloc[:10]:
+        for matchId in matchs["gameId"].iloc[:3]:
             match = getMatch(str(matchId))
             file_path = f"match/{matchId}.json"
+            if os.path.isfile(file_path) == True:
+                break
             with open(file_path, 'w') as outfile:
                 json.dump(match, outfile, indent = 4)
             time.sleep(1.3)
